Fix GPIO read config

+++ mw/gpio_server/gpio_mw.hpp
@@ -34,10 +34,10 @@ class GPIOMWService : public simba::core::ApplicationMW {
  protected:
     std::unique_ptr<com::soc::ISocketStream> sock_;
     std::unique_ptr<core::gpio::IgpioDigitalDriver> gpio_driver_;
-    std::unordered_map<uint16_t, GpioConf> config;
+    std::unordered_map<uint8_t, GpioConf> config;
     std::vector<uint8_t> RxCallback(const std::string& ip, const std::uint16_t& port,
           const std::vector<std::uint8_t> data);
-    std::optional<std::unordered_map<uint16_t, GpioConf>> ReadConfig(nlohmann::json data);
+    std::optional<std::unordered_map<uint8_t, GpioConf>> ReadConfig(nlohmann::json data);
     core::ErrorCode InitPins();
     core::ErrorCode Init(std::unique_ptr<com::soc::ISocketStream> socket,
                               std::unique_ptr<core::gpio::IgpioDigitalDriver> gpio);
